agents-openai/requirements.txt
openai
requests
google-generativeai
python-dotenv